requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
Pillow==10.0.1
Werkzeug==2.3.7 